moved stuff around and linted
