+++ src/nemos/basis/_basis.py
@@ -21,50 +21,6 @@ from ._basis_mixin import BasisTransformerMixin, CompositeBasisMixin
 from ._transformer_basis import TransformerBasis
 
 
-def _recompute_all_default_labels(root: Basis) -> Basis:
-    """Recompute default all labels."""
-    updated = []
-    for bas in root._iterate_over_components():
-        if bas._has_default_label and bas.__class__.__name__ not in updated:
-            bas._recompute_all_labels()
-            updated.append(bas.__class__.__name__)
-    return root
-
-
-def remap_parameters(method):
-    """Remap parameter names to original."""
-
-    @wraps(method)
-    def wrapper(self, **params):
-        # get key/param map
-        _, map_params = self._map_parameters()
-
-        # use mapped key if exists, or original key
-        # deepcopy basis to avoid matching labels when assigning twice the same basis
-        # to two different components:
-        # basis.set_params(basis1=bas, basis2=bas)
-        new_params = {
-            map_params.get(key, key): (
-                copy.deepcopy(val)
-                if isinstance(val, (Basis, TransformerBasis))
-                else val
-            )
-            for key, val in params.items()
-        }
-
-        # apply set_params
-        self = method(self, **new_params)
-
-        # re-assign ordered ids to all default labels of atomic components
-        # note that the recursion always run until parent.
-        if self._parent is None:
-            _recompute_all_default_labels(self)
-
-        return self
-
-    return wrapper
-
-
 def add_docstring(method_name, cls):
     """Prepend super-class docstrings."""
     attr = getattr(cls, method_name, None)
@@ -197,132 +153,6 @@ class Basis(Base, abc.ABC, BasisTransformerMixin):
         # initialize parent to None. This should not end in "_" because it is
         # a permanent property of a basis, defined at composite basis init
         self._parent = None
-
-    def _map_parameters(self, deep=True):
-        """
-        Remap parameters in a given object by replacing 'basis[12]' patterns with unique labels.
-
-        Parameters:
-        -----------
-        bas : object
-            An object that contains a `get_params()` method returning a dictionary of parameters.
-
-        Returns:
-        --------
-        param_dict_map:
-            A dictionary with renamed parameters.
-        key_map:
-            A mapping of new parameter names to original names.
-
-        """
-        param_dict_map, key_map = self._get_params_and_key_map(
-            deep=deep,
-        )  # Retrieve the parameter dictionary
-        # strip higher level label
-        param_dict_map = self._remove_self_label_from_key(param_dict_map)
-        key_map = self._remove_self_label_from_key(key_map)
-        return param_dict_map, key_map
-
-    def __sklearn_get_params__(self, deep=True):
-        """
-        Implements standard scikit-learn get parameters by inspecting init.
-
-        This function will be called by Base.set_params() to get the actual param
-        structure, since the inherited``get_params`` is overridden to use basis labels
-        instead of the nested structure.
-
-        Parameters
-        ----------
-        deep:
-            If true, call itself recursively on basis implementing the method.
-
-        Returns
-        -------
-        out:
-            A dictionary containing the parameters. Key is the parameter
-            name, value is the parameter value.
-        """
-        out = dict()
-        for key in self._get_param_names():
-            value = getattr(self, key)
-            if (
-                deep
-                and hasattr(value, "__sklearn_get_params__")
-                and not isinstance(value, type)
-            ):
-                item_orig = value.__sklearn_get_params__()
-                deep_items = item_orig.items()
-                out.update((key + "__" + k, val) for k, val in deep_items)
-            out[key] = value
-        return out
-
-    def _get_params_and_key_map(self, deep=True) -> Tuple[dict, dict]:
-        """
-        From scikit-learn, get parameters by inspecting init.
-
-        Parameters
-        ----------
-        deep
-
-        Returns
-        -------
-        parameter_dict:
-            A dictionary containing the parameters. Key is the ``basis_label "__" + parameter_name``,
-            value is the parameter value.
-        key_map:
-            Dictionary that maps the keys of parameter_dict onto the keys based on attribute nesting.
-        """
-        parameter_dict = dict()
-        key_map = dict()
-        for key in self._get_param_names():
-            value = getattr(self, key)
-            if (
-                deep
-                and hasattr(value, "_get_params_and_key_map")
-                and not isinstance(value, type)
-            ):
-                item_map, key_mapping = value._get_params_and_key_map()
-                map_deep_items = item_map.items()
-                # only keep the last basis label (the leaf of the tree)
-                parameter_dict.update(
-                    (
-                        (
-                            "__".join(k.split("__")[-2:])
-                            if not isinstance(val, Basis)
-                            else val.label
-                        ),
-                        val,
-                    )
-                    for k, val in map_deep_items
-                )
-                key_map.update(
-                    {
-                        k: key + "__" + v
-                        for k, v in zip(item_map.keys(), key_mapping.values())
-                    }
-                )
-            if isinstance(value, Basis):
-                parameter_dict[value.label] = value
-                key_map[value.label] = key
-            else:
-                parameter_dict[self.label + "__" + key] = value
-                key_map[self.label + "__" + key] = key
-        return parameter_dict, key_map
-
-    def _remove_self_label_from_key(self, map_dict: dict) -> dict:
-        initial_string = self.label + "__"
-        return {
-            k[len(initial_string) :] if k.startswith(initial_string) else k: val
-            for k, val in map_dict.items()
-        }
-
-    def get_params(self, deep=True) -> dict:
-        new_param_dict, _ = self._map_parameters(deep=deep)
-        return new_param_dict
-
-    @remap_parameters
-    def set_params(self, **params: Any):
-        return super().set_params(**params)
 
     @property
     def n_output_features(self) -> int | None:
